requirements.txt
google-api-python-client
pandas
psycopg2-binary
python-dotenv
sqlalchemy
streamlit
plotly